backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
schedule
pandas
numpy
db-sqlite3
python-multipart
# AI library integration if needed (e.g. google-generativeai for direct Gemini use)
google-generativeai
